requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
djangorestframework==3.12.1
pep8==1.7.1
pkg-resources==0.0.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
